Create README.md
The code appears to be a program for controlling a Liquid Crystal Display (LCD) and displaying creative graphics representing a smiling face that blinks, talks, and performs some other movements. The code uses different libraries to control the microcontroller components and display content on the LCD. Some of the used libraries are:

- STD_TYPES: Contains standard type definitions such as uint8_t and uint16_t.
- BIT_MATH: Includes functions for bit manipulation in variables.
- DIO_interface: Used for controlling Digital Input/Output ports.
- LED_interface: Used for controlling LEDs.
- CLCD_interface: Used for controlling the Character LCD (CLCD) display.

The code contains functions to control the facial expressions of the character. These functions include different states for the eyes and mouth, such as smile, close, talk, etc.

There are data arrays in the code that represent the necessary information to draw the facial expressions and movements of the character. These arrays contain the data points that make up the face and its movements.

The code also includes functions like smile(), openEye(), blink(), and others that control the character's facial expressions and movements.

The main loop of the code calls these functions in sequence to execute the character's movements and displays on the LCD.

It's worth noting that this code assumes it is used with a specific Character LCD library and might require modifications to work correctly with other libraries. This code could be an example or experiment for controlling a Character LCD using different libraries.